Log async backtraces after test timeouts
To make it easier to debug in CI why a particular test gets hung.

This is an invasive change, as it requires annotating all async
functions with `async_backtrace::framed`, and using a custom proc macro
(`slatedb_test_macros::test`) rather than the standard `tokio::test`
macro for test functions.